feature(ui-ux): pool swap - slippage tolerance design update (#2200)
* slippage redesign for composite swap screen

* updated testcase for slippage tolerance

* updated test case after removing slippage_select

* Added style for dark mode

* Updated e2e testcase

* added 0.5% slippage tolerance

* ui updates for custom slippage

* auctions details test case update

* fix: linting issue in test

* e2e test update

* Updated e2e for loans
