+++ mobile-app/cypress/integration/functional/wallet/loans/loans.spec.ts
@@ -406,9 +406,7 @@ context('Wallet - Loans - Payback Loans', () => {
     cy.wait(4000)
     cy.getByTestID('text_input_tokenA').type('1').blur()
     cy.wait(3000)
-    cy.getByTestID('slippage_select').click()
     cy.getByTestID('slippage_5%').click()
-    cy.getByTestID('button_tolerance_submit').click()
     cy.getByTestID('button_submit').click()
     cy.getByTestID('button_confirm_swap').click().wait(4000)
     cy.closeOceanInterface()
@@ -442,6 +440,7 @@ context('Wallet - Loans - Payback Loans', () => {
       cy.getByTestID('button_confirm_borrow_loan').click().wait(3000)
       cy.closeOceanInterface()
 
+      cy.wait(3000)
       cy.getByTestID('loans_tabs_YOUR_VAULTS').click()
       cy.getByTestID(`vault_card_${readyVaultIndex}`).click()
       cy.getByTestID('collateral_tab_LOANS').click()
@@ -451,6 +450,7 @@ context('Wallet - Loans - Payback Loans', () => {
       cy.getByTestID('button_confirm_payback_loan').click().wait(4000)
       cy.closeOceanInterface()
 
+      cy.wait(3000)
       cy.getByTestID(`vault_card_${readyVaultIndex}`).click()
       cy.getByTestID('collateral_tab_LOANS').click()
       cy.getByTestID('loan_card_DUSD_payback_loan').click()
@@ -459,7 +459,7 @@ context('Wallet - Loans - Payback Loans', () => {
       cy.getByTestID('payback_input_text').clear().type('102').blur()
       cy.getByTestID('text_resulting_loan_amount').contains('0.00000000')
       cy.getByTestID('text_resulting_col_ratio').contains('N/A')
-      cy.getByTestID('payback_loan_button').click()
+      cy.getByTestID('payback_loan_button').click().wait(3000)
       cy.getByTestID('confirm_title').contains('You are paying')
       cy.getByTestID('text_payment_amount').contains('102.00000000')
       cy.getByTestID('text_payment_amount_suffix').contains('DUSD')
@@ -474,6 +474,7 @@ context('Wallet - Loans - Payback Loans', () => {
       cy.getByTestID('txn_authorization_description')
         .contains('Paying 102.00000000 DUSD')
       cy.closeOceanInterface()
+      cy.wait(3000)
       cy.checkVaultTag('READY', VaultStatus.Ready, `vault_card_${readyVaultIndex}_status`, walletTheme.isDark)
     })
   })
